requirements.txt
python-telegram-bot[job-queue]==21.9
playwright==1.51.0
python-dotenv==1.0.1
